print the command line help if no args given
Prior to this commit, the behavior without arguments is:

```
$ nvtest
==> Error: 'NoneType' object is not callable
```

Which is not very helpful for a newbie who did not know
I had to specify a subcommand to run tests.

After this commit, the behavior is to print the
command line argument help content, which would
inform the user that a subcommand is necessary,
then exit with a nonzero return code.