get a real path when given 'foo.pyt'
If one runs `nvtest run -w foo.pyt`,
currently the "root" in args.paths
is the empty string.
That used to just barely work with
cryptic warnings in previous nvtest
versions, but now fails with the
even more cryptic "Error: no path specified".
This change gives it a proper path.