Bug fix: include vvt files
closes: #47

Parsing of vvt include directives prematurely broke out of the parsing loop, causing
directives that occurred *after* an include file to be skipped.  Additionally, the include
directive was pre-filtered based on on_options and platforms, but that did not allow
filtering based on test name.  The fix was to include all files and combine `when` expressions
so that the included directives could be lazily evaluated with all other directives