Fixed issue with loading and running completed testcases
When a collection of testcases is loaded from the database (say, when running a batch), not only are the individual
testcases loaded, but also their dependencies.  The dependencies should be masked from running, unless they too were
explicitly requested.  But, they were not.  This caused tests to be run unnecessarily when running batches

Closes: #83